Conflicts:
	src/counterexamples/MILPMinimalLabelSetGenerator.h
	src/counterexamples/PathBasedSubsystemGenerator.h
	src/modelchecker/csl/SparseMarkovAutomatonCslModelChecker.h
	src/parser/PrctlParser.cpp
	src/storm.cpp
	test/functional/modelchecker/GmmxxDtmcPrctlModelCheckerTest.cpp
	test/performance/modelchecker/GmmxxDtmcPrctModelCheckerTest.cpp

All conflicts resoved. Revision of src/modelchecker/csl/SparseMarkovAutomatonCslModelChecker.h pending.


Former-commit-id: 6bf9d05dbac6d00b1726c9de11983ebbba47e4ba

+++ src/modelchecker/prctl/SparseDtmcPrctlModelChecker.h
@@ -378,7 +378,7 @@ public:
 
 		// Perform the actual matrix-vector multiplication as long as the bound of the formula is met.
         if (linearEquationSolver != nullptr) {
-            this->linearEquationSolver->performMatrixVectorMultiplication(this->getModel().getTransitionMatrix(), result, &totalRewardVector, formula.getBound());
+			this->linearEquationSolver->performMatrixVectorMultiplication(this->getModel().getTransitionMatrix(), result, &totalRewardVector, static_cast<uint_fast64_t>(formula.getBound()));
         } else {
             throw storm::exceptions::InvalidStateException() << "No valid linear equation solver available.";
         }
